+++ src/components/Accounts/Actions/FeeFineActions.js
@@ -57,12 +57,12 @@ class Actions extends React.Component {
     waives: {
       type: 'okapi',
       records: 'waivers',
-      path: 'waives',
+      path: `waives?limit=${MAX_RECORDS}`,
     },
     refunds: {
       type: 'okapi',
       records: 'refunds',
-      path: 'refunds',
+      path: `refunds?limit=${MAX_RECORDS}`,
     },
     owners: {
       type: 'okapi',
@@ -77,7 +77,7 @@ class Actions extends React.Component {
     transfers: {
       type: 'okapi',
       records: 'transfers',
-      path: 'transfers?limit=2000',
+      path: `transfers?limit=${MAX_RECORDS}`,
     },
     curUserServicePoint: {
       type: 'okapi',
@@ -644,10 +644,12 @@ class Actions extends React.Component {
     const currentFeeFineType = feefines.find(({ feeFineType }) => feeFineType === account?.feeFineType);
     const currentOwnerId = servicePointOwnerId || currentFeeFineType?.ownerId || account?.ownerId;
     const currentOwner = owners.find(o => o.id === currentOwnerId) || {};
+    const accountTypes = this.props.accounts.map(a => a.feeFineType);
+    const checkSomeNotify = feefines.filter(feeFine => feeFine.actionNoticeId && accountTypes.includes(feeFine.feeFineType));
     const initialValues = {
       ownerId: currentOwnerId,
       amount: calculateSelectedAmount(this.props.accounts),
-      notify: !!(currentFeeFineType?.actionNoticeId || currentOwner?.defaultActionNoticeId),
+      notify: actions.transferMany ? checkSomeNotify.length > 0 : !!(currentFeeFineType?.actionNoticeId || currentOwner?.defaultActionNoticeId),
     };
 
     const modals = [