+++ src/components/UserDetailSections/PatronBlock/PatronBlock.js
@@ -20,6 +20,14 @@ import { stripesConnect } from '@folio/stripes/core';
 
 import { calculateSortParams } from '../../util';
 
+import css from './PatronBlock.css';
+
+const PATRON_BLOCKS_COLUMNS = {
+  type: 'type',
+  displayDescription: 'displayDescription',
+  blockedActions: 'blockedActions',
+};
+
 class PatronBlock extends React.Component {
   static manifest = Object.freeze({
     manualPatronBlocks: {
@@ -158,10 +166,24 @@ class PatronBlock extends React.Component {
       },
     } = this.props;
 
+    const pointerWrapper = (content, isManual) => (
+      isManual
+        ? <div className={css.pointerWrapper}>{content}</div>
+        : content
+    );
+
     return {
-      'Type': f => f?.type ?? <FormattedMessage id="ui-users.blocks.columns.automated.type" />,
-      'Display description': f => f.desc || f.message,
-      'Blocked actions': f => {
+      [PATRON_BLOCKS_COLUMNS.type]: f => {
+        const type = f?.type ?? <FormattedMessage id="ui-users.blocks.columns.automated.type" />;
+
+        return pointerWrapper(type, f?.type);
+      },
+      [PATRON_BLOCKS_COLUMNS.displayDescription]: f => {
+        const description = f.desc || f.message;
+
+        return pointerWrapper(description, f?.type);
+      },
+      [PATRON_BLOCKS_COLUMNS.blockedActions]: f => {
         const blockedActions = [];
 
         if (f.borrowing || f.blockBorrowing) {
@@ -176,10 +198,28 @@ class PatronBlock extends React.Component {
           blockedActions.push([formatMessage({ id: 'ui-users.blocks.columns.requests' })]);
         }
 
-        return blockedActions.join(', ');
+        return pointerWrapper(blockedActions.join(', '), f?.type);
       }
     };
   }
+
+  columnMapping = {
+    [PATRON_BLOCKS_COLUMNS.type]: <FormattedMessage id="ui-users.blocks.columns.type" />,
+    [PATRON_BLOCKS_COLUMNS.displayDescription]: <FormattedMessage id="ui-users.blocks.columns.desc" />,
+    [PATRON_BLOCKS_COLUMNS.blockedActions]: <FormattedMessage id="ui-users.blocks.columns.blocked" />,
+  };
+
+  columnWidths = {
+    [PATRON_BLOCKS_COLUMNS.type]: '100px',
+    [PATRON_BLOCKS_COLUMNS.displayDescription]: '350px',
+    [PATRON_BLOCKS_COLUMNS.blockedActions]: '250px',
+  };
+
+  visibleColumns = [
+    PATRON_BLOCKS_COLUMNS.type,
+    PATRON_BLOCKS_COLUMNS.displayDescription,
+    PATRON_BLOCKS_COLUMNS.blockedActions,
+  ];
 
   render() {
     const {
@@ -196,11 +236,6 @@ class PatronBlock extends React.Component {
     } = this.state;
     let contentData = patronBlocks.filter(p => moment(moment(p.expirationDate).endOf('day')).isSameOrAfter(moment().endOf('day')));
     contentData = _.orderBy(contentData, ['metadata.createdDate'], ['desc']);
-    const visibleColumns = [
-      'Type',
-      'Display description',
-      'Blocked actions',
-    ];
 
     const buttonDisabled = this.props.stripes.hasPerm('ui-users.patron_blocks');
     const displayWhenOpen =
@@ -213,16 +248,13 @@ class PatronBlock extends React.Component {
         interactive={false}
         contentData={contentData}
         formatter={this.getPatronFormatter()}
-        visibleColumns={visibleColumns}
+        visibleColumns={this.visibleColumns}
         onHeaderClick={this.onSort}
         sortOrder={sortOrder[0]}
         sortDirection={`${sortDirection[0]}ending`}
         onRowClick={this.onRowClick}
-        columnWidths={{
-          'Type': '100px',
-          'Display description': '350px',
-          'Blocked actions': '250px'
-        }}
+        columnMapping={this.columnMapping}
+        columnWidths={this.columnWidths}
       />;
     const title =
       <Headline size="large" tag="h3">