+++ src/views/UserDetail/UserDetail.js
@@ -7,7 +7,7 @@ import {
   concat,
 } from 'lodash';
 import moment from 'moment';
-import { FormattedMessage } from 'react-intl';
+import { injectIntl, FormattedMessage } from 'react-intl';
 
 import {
   AppIcon,
@@ -63,6 +63,7 @@ import RequestFeeFineBlockButtons from '../../components/RequestFeeFineBlockButt
 import { departmentsShape } from '../../shapes';
 
 import ExportFeesFinesReportButton from './components';
+import ErrorPane from '../../components/ErrorPane';
 
 class UserDetail extends React.Component {
   static propTypes = {
@@ -108,7 +109,7 @@ class UserDetail extends React.Component {
     }).isRequired,
     okapi: PropTypes.shape({
       currentUser: PropTypes.shape({
-        servicePoints: PropTypes.string.isRequired,
+        servicePoints: PropTypes.arrayOf(PropTypes.object).isRequired,
       }).isRequired,
     }).isRequired,
     onClose: PropTypes.func,
@@ -121,6 +122,7 @@ class UserDetail extends React.Component {
     getPreferredServicePoint: PropTypes.func,
     tagsEnabled: PropTypes.bool,
     paneWidth: PropTypes.string,
+    intl: PropTypes.object.isRequired,
   };
 
   static defaultProps = {
@@ -285,6 +287,7 @@ class UserDetail extends React.Component {
   renderDetailsLastMenu(user) {
     const {
       tagsEnabled,
+      intl,
     } = this.props;
 
     const tags = ((user && user.tags) || {}).tagList || [];
@@ -293,17 +296,13 @@ class UserDetail extends React.Component {
       <PaneMenu>
         {
           tagsEnabled &&
-          <FormattedMessage id="ui-users.showTags">
-            {ariaLabel => (
-              <IconButton
-                icon="tag"
-                id="clickable-show-tags"
-                onClick={() => { this.showHelperApp('tags'); }}
-                badgeCount={tags.length}
-                aria-label={ariaLabel}
-              />
-            )}
-          </FormattedMessage>
+            <IconButton
+              icon="tag"
+              id="clickable-show-tags"
+              onClick={() => { this.showHelperApp('tags'); }}
+              badgeCount={tags.length}
+              aria-label={intl.formatMessage({ id: 'ui-users.showTags' })}
+            />
         }
       </PaneMenu>
     );
@@ -341,11 +340,13 @@ class UserDetail extends React.Component {
     if (showActionMenu) {
       return (
         <>
-          <RequestFeeFineBlockButtons
-            barcode={barcode}
-            onToggle={onToggle}
-            userId={this.props.match.params.id}
-          />
+          <IfInterface name="feesfines">
+            <RequestFeeFineBlockButtons
+              barcode={barcode}
+              onToggle={onToggle}
+              userId={this.props.match.params.id}
+            />
+          </IfInterface>
           <IfPermission perm="ui-users.edit">
             <Button
               buttonStyle="dropdownItem"
@@ -362,11 +363,13 @@ class UserDetail extends React.Component {
               </Icon>
             </Button>
           </IfPermission>
-          <ExportFeesFinesReportButton
-            feesFinesReportData={feesFinesReportData}
-            onToggle={onToggle}
-            callout={this.callout}
-          />
+          <IfInterface name="feesfines">
+            <ExportFeesFinesReportButton
+              feesFinesReportData={feesFinesReportData}
+              onToggle={onToggle}
+              callout={this.callout}
+            />
+          </IfInterface>
         </>
       );
     } else {
@@ -405,6 +408,10 @@ class UserDetail extends React.Component {
 
       return { ...address, addressType };
     });
+  }
+
+  userNotFound = () => {
+    return this.props.resources?.selUser?.failed?.httpStatus === 404;
   }
 
   render() {
@@ -453,11 +460,25 @@ class UserDetail extends React.Component {
       'addressType',
       '',
     );
-    const customFields = user?.customFields || [];
+    const customFields = user?.customFields || {};
     const departments = resources?.departments?.records || [];
     const userDepartments = (user?.departments || [])
       .map(departmentId => departments.find(({ id }) => id === departmentId)?.name);
     const accounts = resources?.accounts;
+
+    if (this.userNotFound()) {
+      return (
+        <ErrorPane
+          id="pane-user-not-found"
+          defaultWidth={paneWidth}
+          paneTitle={<FormattedMessage id="ui-users.information.userDetails" />}
+          dismissible
+          onClose={this.onClose}
+        >
+          <FormattedMessage id="ui-users.errors.userNotFound" />
+        </ErrorPane>
+      );
+    }
 
     if (!user) {
       return (
@@ -673,4 +694,4 @@ class UserDetail extends React.Component {
   }
 }
 
-export default UserDetail;
+export default injectIntl(UserDetail);