+++ CHANGELOG.md
@@ -1,5 +1,38 @@
 # Change history for ui-users
 
+## 6.1.0 (IN PROGRESS)
+
+* Configure Jest/RTL. Refs UIU-2112.
+* Show user-readable message when user is not found. Fixes UIU-2081.
+* Fix Custom Fields error message by adding a missing permission. Fixes UIU-2104.
+* Add changes to indicate clickable for cursor. Refs UIU-2052.
+* Migrate from string notation to column mapping for PatronBlock. Refs UIU-2091.
+* Add Cash Drawer Reconciliaton report & Financial transactions detail report permissions to reports. Refs UIU-2088.
+* Add new fee/fine reports as options to User Actions drop-down. Refs UIU-2083.
+* Show claimed returned report only if user has necessary permission. Refs UIU-1884.
+* Fix Note Edit Page - doesn't return to record page after click on `Save&Close`. Fixes UIU-2087.
+* Fix Notify patron box behavior for New fee/fine when default patron notice is set. Refs UIU-2111.
+* Close "New fee/fine" page after fee/fine created. Refs UIU-2117.
+* Fix disabling "Save & close" button for Manual patron block. Refs UIU-2123.
+* Fix the possibility of create manual patron block with expiration date of today. Refs UIU-2122.
+* Fix permission error for "Refunds to Process Manually" report. Refs UIU-2126.
+* Also support `circulation` `10.0`. Refs UIU-2135.
+* Show the "Refunds to process manually" report conditionally based on permissions. Refs UIU-2125.
+* Fetch some search container routes conditionally based on permissions. Refs UIU-2132.
+* Handle search of ASTified translation values. Refs UIU-2142.
+* Fix optional dependencies to be actually optional and add a few. UIU-2140.
+* Include missing `limit` clause in request-count query. Refs UIU-2143.
+* Replace local KeyboardShortcutsModal component with import. Refs UIU-2151.
+* Clean up prop-types that generate bogus console warnings. Refs UIU-2158.
+* Provide useful `aria-label` values to loan action (ellipses) menues. Refs UIU-1635.
+* Allow a user to assign an existing permission set to a permission set. Refs UIU-1630.
+* Add permissions type filter to users settings permissions sets. Refs UIU-2164.
+* Avoid querying for an empty list of loan policies. Refs UIU-2163.
+* Show `Pop-up on User` and `Pop-up on Checkout` fields to Notes forms. Refs UIU-2155.
+* Show loading indicator on Loans details screen. Fixes UIU-2120.
+* Add type column to users settings permissions sets. Refs UIU-2167.
+* Add modal for 'Cash drawer reconciliation report'. Refs UIU-1959.
+
 ## [6.0.0](https://github.com/folio-org/ui-users/tree/v6.0.0) (2021-03-18)
 [Full Changelog](https://github.com/folio-org/ui-users/compare/v5.0.9...v6.0.0)
 
@@ -76,6 +109,7 @@
 * Fix eslint error in `LoanDetails.js`. Refs UIU-2068.
 * Improve fetching account data by making sure fetch happens only once. Fixes UIU-2063.
 * Update `@folio/plugin-find-user` for compatibility with `@folio/stripes` `v6`.
+* Fix behavior of Confirm button when `Mark as missing` option is selected in `Resolve claim` menu. Fixes UIU-2077.
 
 ## [5.0.9](https://github.com/folio-org/ui-users/tree/v5.0.9) (2020-12-10)
 [Full Changelog](https://github.com/folio-org/ui-users/compare/v5.0.8...v5.0.9)